requirements.txt
matplotlib==3.3.4
cryptography==3.4.7
pandas==1.1.5
phe==1.4.0
pycryptodome==3.10.1
